[Refactor]James - Finished refactoring and modularizing the code to make it more organized and managable
